+++ sync-for-payables/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v1.3.0] sync-for-payables
 ### Releases
 - [NPM v1.3.0] https://www.npmjs.com/package/@codat/sync-for-payables/v/1.3.0 - sync-for-payables
+
+## 2023-10-16 01:23:52
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.99.2 (2.155.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.3.1] sync-for-payables
+### Releases
+- [NPM v1.3.1] https://www.npmjs.com/package/@codat/sync-for-payables/v/1.3.1 - sync-for-payables